src/third_party/blink/web_tests/virtual/layout-ng-grid/external/wpt/css/css-grid/alignment/grid-place-content-001-expected.txt

This is a testharness.js-based test.
PASS .grid 1
FAIL .grid 2 assert_equals: 
<div class="grid small content-box center">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 40
FAIL .grid 3 assert_equals: 
<div class="grid small content-box end">
  <div data-offset-x="0" data-offset-y="0">X</div>
</div>
offsetLeft expected 0 but got 40
PASS .grid 4
FAIL .grid 5 assert_equals: 
<div class="grid small content-box space-around">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 40
PASS .grid 6
FAIL .grid 7 assert_equals: 
<div class="grid small border-box center">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 40
FAIL .grid 8 assert_equals: 
<div class="grid small border-box end">
  <div data-offset-x="0" data-offset-y="0">X</div>
</div>
offsetLeft expected 0 but got 40
PASS .grid 9
FAIL .grid 10 assert_equals: 
<div class="grid small border-box space-around">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 40
PASS .grid 11
FAIL .grid 12 assert_equals: 
<div class="grid big content-box center">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 0
FAIL .grid 13 assert_equals: 
<div class="grid big content-box end">
  <div data-offset-x="40" data-offset-y="40">X</div>
</div>
offsetLeft expected 40 but got 0
PASS .grid 14
FAIL .grid 15 assert_equals: 
<div class="grid big content-box space-around">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 0
PASS .grid 16
FAIL .grid 17 assert_equals: 
<div class="grid big border-box center">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 0
FAIL .grid 18 assert_equals: 
<div class="grid big border-box end">
  <div data-offset-x="40" data-offset-y="40">X</div>
</div>
offsetLeft expected 40 but got 0
PASS .grid 19
FAIL .grid 20 assert_equals: 
<div class="grid big border-box space-around">
  <div data-offset-x="20" data-offset-y="20">X</div>
</div>
offsetLeft expected 20 but got 0
Harness: the test ran to completion.
